Refactor block list generation
