+++ test_modules/get_webpage_tests.py
@@ -16,7 +16,8 @@
 
 import unittest
 
-class TestWebScraper(unittest.TestCase):
+
+class TestGetWebpage(unittest.TestCase):
 
     def setUp(self):
         pass
